WIP sidecar refactor
