+++ internal/app/sidecar/constants/constants.go
@@ -1,0 +1,27 @@
+package constants
+
+const (
+	// Prepare indicates a prepare action
+	Prepare = "prepare"
+
+	// Commit indicates a commit action
+	Commit = "commit"
+
+	// DevBaseDir is the base directory for dev output
+	DevBaseDir = ".dev"
+
+	// InputBlobDir is the input blob data directory
+	InputBlobDir = "in/data"
+
+	// InputMetaFile is the input meta data file
+	InputMetaFile = "in/meta.json"
+
+	// OutputBlobDir is the output blob data directory
+	OutputBlobDir = "out/data"
+
+	// OutputMetaFile is the output meta data file
+	OutputMetaFile = "out/meta.json"
+
+	// OutputEventsDir is the output event data
+	OutputEventsDir = "out/events"
+)
