+++ internal/app/sidecar/integration/dev_integration_test.go
@@ -3,20 +3,19 @@ package integration
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/lawrencegripper/ion/internal/app/sidecar/module"
 	"io/ioutil"
-	"net/http"
 	"os"
 	"path"
-	"runtime"
+	"path/filepath"
+	"strings"
+
 	"testing"
 
-	"github.com/lawrencegripper/ion/internal/app/sidecar/app"
-	"github.com/lawrencegripper/ion/internal/app/sidecar/blob/filesystem"
-	"github.com/lawrencegripper/ion/internal/app/sidecar/events/mock"
-	"github.com/lawrencegripper/ion/internal/app/sidecar/meta/inmemory"
-	"github.com/lawrencegripper/ion/internal/app/sidecar/types"
+	"github.com/lawrencegripper/ion/internal/app/sidecar"
+	"github.com/lawrencegripper/ion/internal/app/sidecar/constants"
+	"github.com/lawrencegripper/ion/internal/app/sidecar/dataplane/metadata/inmemory"
 	"github.com/lawrencegripper/ion/internal/pkg/common"
-	"github.com/sirupsen/logrus"
 )
 
 var sharedDB *inmemory.InMemoryDB
@@ -27,61 +26,60 @@ func TestDevIntegration(t *testing.T) {
 		t.Skip("Skipping integration test in short mode...")
 	}
 
-	// Setting the base directory to empty
-	// will result in /ion/... being used.
-	baseDir := ""
-	if runtime.GOOS == "windows" {
-		// Use a relative base directory
-		// on Windows to avoid Administrator
-		// issues.
-		baseDir = "ion"
-	}
-	outDir := path.Join(baseDir, "out")
-	outDataDir := path.Join(outDir, "data")
-	outMetaFilePath := path.Join(outDir, "meta.json")
-	outEventsDir := path.Join(outDir, "events")
-	outEventFilePath := path.Join(outEventsDir, "event1.json")
-	inDir := path.Join(baseDir, "in")
-	inDataDir := path.Join(inDir, "data")
-	inMetaFilePath := path.Join(inDir, "meta.json")
+	// Create context
 	eventID := "1111111"
-
-	config := &app.Configuration{
-		SharedSecret: "secret",
-		BaseDir:      baseDir,
-		Context: &common.Context{
-			Name:          "testmodule",
-			EventID:       eventID,
-			CorrelationID: "fish",
-			ParentEventID: "",
-		},
-		ServerPort:  8080,
-		PrintConfig: false,
-		LogLevel:    "Debug",
-		Development: true,
+	baseDir := "ion"
+	eventTypesStr := "face_detected"
+	eventTypes := strings.Split(eventTypesStr, ",")
+	context := &common.Context{
+		Name:          "testmodule",
+		EventID:       eventID,
+		CorrelationID: "fish",
+		ParentEventID: "",
 	}
 
-	module1, err := createDevModule(config)
+	// Configuration for module 1
+	config := sidecar.Configuration{
+		Action:          constants.Prepare,
+		BaseDir:         baseDir,
+		Context:         context,
+		ValidEventTypes: eventTypesStr,
+		PrintConfig:     false,
+		LogLevel:        "Debug",
+		Development:     true,
+	}
+
+	environment := module.GetModuleEnvironment(baseDir)
+
+	err := sidecar.Run(config)
 	if err != nil {
-		t.Error(err)
+		t.Fatal(err)
 	}
-	defer module1.Close() // This is to ensure cleanup
+
 	defer func() {
-		_ = os.RemoveAll(types.DevBaseDir) // Clean up development files
+		_ = os.RemoveAll(baseDir)
+		_ = os.RemoveAll(constants.DevBaseDir)
+		_ = os.Remove(".memdb")
 	}()
+
+	// Check dev.prepared exists in development dir
+	preparedPath := filepath.FromSlash(path.Join(constants.DevBaseDir, eventID, "dev.prepared"))
+	if _, err := os.Stat(preparedPath); os.IsNotExist(err) {
+		t.Errorf("dev.prepared file should exist at path '%s'", preparedPath)
+	}
 
 	// Write an output image blob
 	blob1 := "img1.png"
-	blob1FilePath := path.Join(outDataDir, blob1)
+	blob1FilePath := filepath.FromSlash(path.Join(environment.OutputBlobDirPath, blob1))
 	writeOutputBlob(blob1FilePath)
 
 	// Write an output image blob
 	blob2 := "img2.png"
-	blob2FilePath := path.Join(outDataDir, blob2)
+	blob2FilePath := filepath.FromSlash(path.Join(environment.OutputBlobDirPath, blob2))
 	writeOutputBlob(blob2FilePath)
 
 	// Grab the length of the output directory
-	outFiles, err := ioutil.ReadDir(outDataDir)
+	outFiles, err := ioutil.ReadDir(environment.OutputBlobDirPath)
 	if err != nil {
 		t.Errorf("error reading out dir '%+v'", err)
 	}
@@ -89,134 +87,93 @@ func TestDevIntegration(t *testing.T) {
 
 	// Write an output metadata file
 	insight := []byte(`[{"key": "key2","value": "value2"}]`)
-	writeOutputBytes(insight, outMetaFilePath)
+	writeOutputBytes(insight, environment.OutputMetaFilePath)
 
 	// Write an output event file
 	j := fmt.Sprintf(`[{"key":"eventType","value":"%s"},{"key":"files","value":"%s,%s"},{"key":"abc","value":"123"}]`, eventTypes[0], blob1, blob2)
 	outEvent := []byte(j)
 
-	writeOutputBytes(outEvent, outEventFilePath)
+	writeOutputBytes(outEvent, filepath.FromSlash(path.Join(environment.OutputEventsDirPath, "event1.json")))
 
-	client := &http.Client{}
-
-	// Ready will attempt to sync the execution environment for this module - this should be empty
-	if err := executeRequest(client, config.SharedSecret, fmt.Sprintf("%v", config.ServerPort), "ready"); err != nil {
-		t.Errorf("error calling ready '%+v'", err)
-	}
-
-	// Check dev.ready exists in development dir
-	readyPath := path.Join(types.DevBaseDir, eventID, "dev.ready")
-	if _, err := os.Stat(readyPath); os.IsNotExist(err) {
-		t.Errorf("dev.ready file should exist at path '%s'", readyPath)
-	}
-
-	// Done will commit the written files to external providers
-	if err := executeRequest(client, config.SharedSecret, fmt.Sprintf("%v", config.ServerPort), "done"); err != nil {
-		t.Errorf("error calling done '%+v'", err)
-	}
-
-	// Check dev.done exists in development dir
-	donePath := path.Join(types.DevBaseDir, eventID, "dev.done")
-	if _, err := os.Stat(donePath); os.IsNotExist(err) {
-		t.Errorf("dev.done file should exist at path '%s'", donePath)
-	}
-
-	// Clear local module directories
-	module1.Close()
-
-	// Hydrate event
-	b, err := ioutil.ReadFile(path.Join(types.DevBaseDir, config.Context.EventID, "events", "event0.json"))
+	config.Action = constants.Commit
+	err = sidecar.Run(config)
 	if err != nil {
-		t.Errorf("error reading event from disk '%+v'", err)
+		t.Fatal(err)
+	}
+
+	// Check dev.committed exists in development dir
+	committedPath := filepath.FromSlash(path.Join(constants.DevBaseDir, eventID, "dev.committed"))
+	if _, err := os.Stat(committedPath); os.IsNotExist(err) {
+		t.Fatalf("dev.committed file should exist at path '%s'", committedPath)
+	}
+
+	// Grab event ID from module 1's output event
+	b, err := ioutil.ReadFile(filepath.FromSlash(path.Join(constants.DevBaseDir, eventID, "events", "event0.json")))
+	if err != nil {
+		t.Fatalf("error reading event from disk '%+v'", err)
 	}
 	var inEvent common.Event
 	err = json.Unmarshal(b, &inEvent)
 	if err != nil {
-		t.Errorf("error unmarshalling event '%+v'", err)
+		t.Fatalf("error unmarshalling event '%+v'", err)
 	}
 
+	// Set module 1 as the parent and set the input event ID
 	config.Context.ParentEventID = config.Context.EventID
 	config.Context.EventID = inEvent.Context.EventID
-	module2, err := createDevModule(config)
+	config.Action = constants.Prepare
+	err = sidecar.Run(config)
 	if err != nil {
-		t.Error(err)
-	}
-	defer module2.Close()
-
-	// Ready will attempt to sync the execution environment for this module.
-	// This should download the files written by the previous done.
-	if err := executeRequest(client, config.SharedSecret, fmt.Sprintf("%v", config.ServerPort), "ready"); err != nil {
-		t.Errorf("error calling done '%+v'", err)
+		t.Fatal(err)
 	}
 
 	// Check blob input data matches the output from the first module
-	inFiles, err := ioutil.ReadDir(inDataDir)
+	inFiles, err := ioutil.ReadDir(environment.InputBlobDirPath)
 	if err != nil {
-		t.Errorf("error reading in dir '%+v'", err)
+		t.Fatalf("error reading in dir '%+v'", err)
 	}
 	inLength := len(inFiles)
 
 	if (inLength != outLength) && outLength > 0 {
-		t.Errorf("error, input files length should match output length")
+		t.Fatalf("error, input files length should match output length")
 	}
 
 	// Check the input metadata is the same as that output from the first module
-	inMetaData, err := ioutil.ReadFile(inMetaFilePath)
+	inMetaData, err := ioutil.ReadFile(environment.InputMetaFilePath)
 	if err != nil {
-		t.Errorf("error reading in meta file '%s': '%+v'", inMetaFilePath, err)
+		t.Fatalf("error reading in meta file '%s': '%+v'", environment.InputMetaFilePath, err)
 	}
 
 	var kvps common.KeyValuePairs
 	err = json.Unmarshal(inMetaData, &kvps)
 	if err != nil {
-		t.Errorf("error decoding file '%s' content: '%+v'", inMetaFilePath, err)
+		t.Fatalf("error decoding file '%s' content: '%+v'", environment.InputMetaFilePath, err)
 	}
 
 	// The first key, value pair should be as expected
 	for _, kvp := range kvps {
 		if kvp.Key != "abc" {
-			t.Errorf("expected key 'abc' in key value pairs: '%+v'", kvp)
+			t.Fatalf("expected key 'abc' in key value pairs: '%+v'", kvp)
 		}
 		if kvp.Value != "123" {
-			t.Errorf("expected key 'abc' to have value '123' in key value pairs: '%+v'", kvp)
+			t.Fatalf("expected key 'abc' to have value '123' in key value pairs: '%+v'", kvp)
 		}
 		break
 	}
 }
 
-func createDevModule(config *app.Configuration) (*app.App, error) {
-	if sharedDB == nil {
-		var err error
-		sharedDB, err = inmemory.NewInMemoryDB()
-		if err != nil {
-			return nil, fmt.Errorf("Failed to establish metadata store with debug provider, error: %+v", err)
-		}
-	}
-	fsBlob, err := filesystem.NewBlobStorage(&filesystem.Config{
-		InputDir:  path.Join(types.DevBaseDir, config.Context.ParentEventID, "blobs"),
-		OutputDir: path.Join(types.DevBaseDir, config.Context.EventID, "blobs"),
-	})
+func writeOutputBlob(path string) error {
+	err := ioutil.WriteFile(path, []byte("image1"), 0777)
 	if err != nil {
-		return nil, fmt.Errorf("Failed to establish metadata store with debug provider, error: %+v", err)
+		return fmt.Errorf("error writing file '%s', '%+v'", path, err)
 	}
-	fsEvents := mock.NewEventPublisher(path.Join(types.DevBaseDir, config.Context.EventID, "events"))
+	return nil
+}
 
-	logger := logrus.New()
-	logger.Out = os.Stdout
-
-	a := app.App{}
-	a.Setup(
-		config.SharedSecret,
-		config.BaseDir,
-		config.Context,
-		eventTypes,
-		sharedDB,
-		fsEvents,
-		fsBlob,
-		logger,
-		config.Development,
-	)
-
-	go a.Run(fmt.Sprintf(":%d", config.ServerPort))
-	return &a, nil
+func writeOutputBytes(bytes []byte, path string) error {
+	err := ioutil.WriteFile(path, bytes, 0777)
+	if err != nil {
+		return fmt.Errorf("error writing file '%s', '%+v'", bytes, err)
+	}
+	return nil
 }
